Backgrounding and foregrounding file system commands works infinity times.
* When fork/exec'ing have the forked process become a session leader
* Send SIGSTOP signal when suspending to child process
* Send SIGCONT signal when resuming a suspended child process
* Send SIGINT signal when interrupting a child process

Becoming a session leader is important because w/o it any signal sent to the parent process (lagniappe) would automatically have been sent. If you have more than a suspended process hitting Ctrl-C to send an interrupt would kill both the foreground and the suspended process which is not what a user would expect. A user would expect the foreground process to stop, but the suspended process should be left alone.