Recursively expand aliases before processing commands.
e.g.

   alias foo='ls | !upcase | grep RB'
   alias bar='foo | !downcase | grep rb'
   alias baz='bar | !to_s'
   baz

When running 'baz' it expands the bar reference out to the contents of the bar alias. Then it will process the process alias and expand the foo reference out to the foo alias. Then it will look to see if the foo alias should expand, but there are no aliases referenced so it stops.

The result is:
    ls | !upcase | grep rb | !downcase | grep rb | !to_s