Resolving suspend/resume (SIGTSTP/SIGCONT) issue.
The forked child process needs to be put into its
own process group that it can become the leader of.

BRAIN DUMP
----------

The TTY needs the shell and its child processes to
belong to different process groups and then the control
of the TTY should be given to whichever process group
should be in the foreground (only one at at time).

This resolves the issue where the TTY input buffer
got confused when the shell and its children belonged
to the same process group and a SIGTSTP was sent.

Much thanks to kurahaupo in #bash on irc.freenode.net
for offering up a theory about this being related to
process groups.

ADDITIONAL NOTES
----------------
The shell no longer listens for any signals that it
should forward to its child processes. It ignores all
of the signals it doesn't care about and then uses
Process.wait2 to wait for a child to stop (suspend or exit).

When forking a child process reset all of the signals
that may be ignored.