Etching in a loading mechanism for sandboxing addons from one another.
ADDON DIRECTORY
----------------

Every add-on gets a unique directory name. The uniqueness here is to allow new add-ons to be added without namespace pollution.


ADDON FILES
------------

Every add-on directory can have any number of files below it. These files can be organized however the add-on wants them to be with the following caveats:

   * top-level add-on files will be loaded by Yap (load order is not guaranteed)
   * support-files for add-ons can be nested inside of the lib/ directory and will be available to load (see next commit).