Updating pre_word_context for tab completion.
pre_word_context is the text that comes before the current word based on where the cursor is. Here's an example: user types "ls lib/" and then hits tab. 

The current word as reported by RawLine is "" because the forward-slash is a word break character. In order to complete correctly we need set "lib/" to the pre_word_context.