Pass around the current instance of world so it can be used to lookup the current environment among other things.
  * currently just focused on env variables, no other aspect of world are being used

  * be sure to set the ENV just before executing a file system command and to restore after-wards. This works around an issue where even if you pass in an env hash to Kernel.exec it would still inherit missing env variables from the current ENV. 

  * the shell's env starts out inheriting from the parent process. This may change in the future.

MORE INFO:

This resolves the issue where the env vars for Yap were competing with the env vars for the user's current shell. Those two things should never conflict in order to ensure a stable yap instance and to also allow for the user to have full authority over the env variables used for their shell.