Updating tab completion to start by showing the common-most string amongst all possible matches.
For example, say you have "Gemfile" and "Gemfile.lock" in a directory and you type "ls g" and hit tab. This will now auto-expand to show "Gemfile" as the common-string: 

   ls Gemfile

This lets the user hit <enter> if that's what they wanted or continue to hit tab to cycle thru other matches. It's one less step for the user.