Shell functions can now be defined dynamically.
* Shell functions no longer need to specify all incoming block parameters. Only specify what you need.
* Shell functions no longer are limited to just static names. They can now be dynamically matched. The registered function object only needs to respond/implement #match(...)

For example, say you have a dot function which is a convenience method for going up any number of directory levels. So ".." goes up one directory, "..." goes to two directories, "...." goes up three directories, and so on. 

Previously, you couldn't easily define this without having an individual alias or function for each set of dots. Now, you can:

    func /^\.{2,}$/ do |command:|
      (command.length - 1).times do
        Dir.chdir("..")
      end
    end


Thanks to @brandondrew for this suggestion and initial thoughts on implementation