Provide the line of input being processed to commands executing it.
This allows user-defined functions to be able to operate on the line itself if it's command matches a shell command.

Introduces World#shell which will allows user-defined functions to pass input thru and have it evaluated by yap.

Lastly, providing an example in yaprc for implementing loops/repetition, e.g.:

   10.times: echo hi
   10.times as n: echo hi $n
   0..9: echo foo
   0..9 as i: echo this is $i