Poor man's fuzzy searching backtracking bug fix.
I'm not entirely certain this is correct, but I thought it was clever. 

The problem is that on long input strings an explosion of backtracking may occur. When this occurs it will appear as if the shell hung, but eventually it will finish.

To work around this issue append whatever it is being searched for the end of the input line. This guarantees a match. However, if we match after a position greater than the length of the original item then we know we failed.

Let's go with it for now.